test_force.js has a new option --no-plugin-update
When provided, we don't run ./tools/update.sh on the clone of the plugin repo when testing forcehybrid
Using that new option in CI to speed testing up.

In the past, cordova plugin repo had binaries, and we only updated them on release day.
Now it only hold sources, so as long as the sources are up to date, there is no point running ./tools/update.sh when testing hybrid.