New hidden arg (sdkdependencies) for forceios/forcedroid/etc to override sdk dependencies. It's also exposed as a new optional argument for test_force.js.
Use SDK dependencies override if you need to generate/build a template pointing to some fork/branch of Mobile SDK.
To do that before, you would have to change the package.json of the template in your fork, check it in and then run test_force.js with templaterepouri pointing to your fork.

Example:
To build the template using the forked Salesforce Mobile SDK iOS repo
Pass the following:
--sdkdependencies={\"SalesforceMobileSDK-iOS\":\"https://github.com/somefork/SalesforceMobileSDK-iOS#somebranch\"}

There is also a new flag spm-update for test_force.js
When provided the SPM repo is cloned and build locally and then the CLIs are invoked with SalesforceMobileSDK-iOS-SPM pointing to the local directory (using the sdk dependencies override just described).
That way we can make sure that our SPM template generates and builds with the latest iOS code without having to build and checkin the xcframeworks (which we only do at release time).